Move samples for `drop`, `dropNulls`, `dropNaNs`, `dropNA` to the `samples` module
