+++ docs/StardustDocs/topics/_shadow_resources.md
@@ -235,10 +235,6 @@
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.distinct.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.distinctBy.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.distinctColumns.html"></resource>
-<resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.dropNA.html"></resource>
-<resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.dropNaNs.html"></resource>
-<resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.dropNulls.html"></resource>
-<resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.dropWhere.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.filter.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.getColumnsByName.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.getRowByCondition.html"></resource>
@@ -362,3 +358,19 @@
 <resource src="takeLastDataColumn.html"></resource>
 <resource src="dropDataColumn.html"></resource>
 <resource src="dropLastDataColumn.html"></resource>
+<resource src="dropWhere_properties.html"></resource>
+<resource src="dropNulls.html"></resource>
+<resource src="dropNullsWhereAllNull.html"></resource>
+<resource src="dropNullsSelector.html"></resource>
+<resource src="dropNullsSelectorSeveralCols.html"></resource>
+<resource src="dropNullsWhereAllNullSelector.html"></resource>
+<resource src="dropNaNs.html"></resource>
+<resource src="dropNaNsWhereAllNaN.html"></resource>
+<resource src="dropNaNsSelector.html"></resource>
+<resource src="dropNaNsSelectorSeveralCols.html"></resource>
+<resource src="dropNaNsWhereAllNaNSelector.html"></resource>
+<resource src="dropNA.html"></resource>
+<resource src="dropNAWhereAllNA.html"></resource>
+<resource src="dropNASelector.html"></resource>
+<resource src="dropNASelectorSeveralCols.html"></resource>
+<resource src="dropNAWhereAllNASelector.html"></resource>
