+++ docs/StardustDocs/topics/drop.md
@@ -1,6 +1,6 @@
 [//]: # (title: drop / dropNulls / dropNaNs / dropNA)
 
-<!---IMPORT org.jetbrains.kotlinx.dataframe.samples.api.Access-->
+<!---IMPORT org.jetbrains.kotlinx.dataframe.samples.api.DropNullsNansNa-->
 
 Removes all rows that satisfy [row condition](DataRow.md#row-conditions)
 
@@ -22,8 +22,8 @@ df.drop { it["weight"] == null || it["city"] == null }
 ```
 
 </tab></tabs>
-<inline-frame src="resources/org.jetbrains.kotlinx.dataframe.samples.api.Access.dropWhere.html" width="100%"/>
 <!---END-->
+<inline-frame src="./resources/dropWhere_properties.html" width="100%" height="500px"></inline-frame>
 
 If called on a [`DataColumn`](DataColumn.md), removes all the values that match the predicate 
 and returns a [`DataColumn`](DataColumn.md) containing the values that do not match the predicate.
@@ -38,14 +38,46 @@ See [column selectors](ColumnSelectors.md) for how to select the columns for thi
 
 ```kotlin
 df.dropNulls() // remove rows with null value in any column
+```
+
+<!---END-->
+<inline-frame src="./resources/dropNulls.html" width="100%" height="500px"></inline-frame>
+
+<!---FUN dropNullsWhereAllNull-->
+
+```kotlin
 df.dropNulls(whereAllNull = true) // remove rows with null values in all columns
+```
+
+<!---END-->
+<inline-frame src="./resources/dropNullsWhereAllNull.html" width="100%" height="500px"></inline-frame>
+
+<!---FUN dropNullsSelector-->
+
+```kotlin
 df.dropNulls { city } // remove rows with null value in 'city' column
+```
+
+<!---END-->
+<inline-frame src="./resources/dropNullsSelector.html" width="100%" height="500px"></inline-frame>
+
+<!---FUN dropNullsSelectorSeveralCols-->
+
+```kotlin
 df.dropNulls { city and weight } // remove rows with null value in 'city' OR 'weight' columns
+```
+
+<!---END-->
+<inline-frame src="./resources/dropNullsSelectorSeveralCols.html" width="100%" height="500px"></inline-frame>
+
+<!---FUN dropNullsWhereAllNullSelector-->
+
+```kotlin
 df.dropNulls(whereAllNull = true) { city and weight } // remove rows with null value in 'city' AND 'weight' columns
 ```
 
-<inline-frame src="resources/org.jetbrains.kotlinx.dataframe.samples.api.Access.dropNulls.html" width="100%"/>
 <!---END-->
+<inline-frame src="./resources/dropNullsWhereAllNullSelector.html" width="100%" height="500px"></inline-frame>
 
 If called on a [`DataColumn`](DataColumn.md), removes `null` values from this [`DataColumn`](DataColumn.md), 
 adjusting the type accordingly.
@@ -60,14 +92,46 @@ See [column selectors](ColumnSelectors.md) for how to select the columns for thi
 
 ```kotlin
 df.dropNaNs() // remove rows containing NaN in any column
+```
+
+<!---END-->
+<inline-frame src="./resources/dropNaNs.html" width="100%" height="500px"></inline-frame>
+
+<!---FUN dropNaNsWhereAllNaN-->
+
+```kotlin
 df.dropNaNs(whereAllNaN = true) // remove rows with NaN in all columns
+```
+
+<!---END-->
+<inline-frame src="./resources/dropNaNsWhereAllNaN.html" width="100%" height="500px"></inline-frame>
+
+<!---FUN dropNaNsSelector-->
+
+```kotlin
 df.dropNaNs { weight } // remove rows where 'weight' is NaN
+```
+
+<!---END-->
+<inline-frame src="./resources/dropNaNsSelector.html" width="100%" height="500px"></inline-frame>
+
+<!---FUN dropNaNsSelectorSeveralCols-->
+
+```kotlin
 df.dropNaNs { age and weight } // remove rows where either 'age' or 'weight' is NaN
+```
+
+<!---END-->
+<inline-frame src="./resources/dropNaNsSelectorSeveralCols.html" width="100%" height="500px"></inline-frame>
+
+<!---FUN dropNaNsWhereAllNaNSelector-->
+
+```kotlin
 df.dropNaNs(whereAllNaN = true) { age and weight } // remove rows where both 'age' and 'weight' are NaN
 ```
 
-<inline-frame src="resources/org.jetbrains.kotlinx.dataframe.samples.api.Access.dropNaNs.html" width="100%"/>
 <!---END-->
+<inline-frame src="./resources/dropNaNsWhereAllNaNSelector.html" width="100%" height="500px"></inline-frame>
 
 If called on a [`DataColumn`](DataColumn.md), removes [`NaN` values](nanAndNa.md#nan) from this [`DataColumn`](DataColumn.md),
 adjusting the type accordingly.
@@ -82,14 +146,46 @@ See [column selectors](ColumnSelectors.md) for how to select the columns for thi
 
 ```kotlin
 df.dropNA() // remove rows containing null or NaN in any column
+```
+
+<!---END-->
+<inline-frame src="./resources/dropNA.html" width="100%" height="500px"></inline-frame>
+
+<!---FUN dropNAWhereAllNA-->
+
+```kotlin
 df.dropNA(whereAllNA = true) // remove rows with null or NaN in all columns
+```
+
+<!---END-->
+<inline-frame src="./resources/dropNAWhereAllNA.html" width="100%" height="500px"></inline-frame>
+
+<!---FUN dropNASelector-->
+
+```kotlin
 df.dropNA { weight } // remove rows where 'weight' is null or NaN
+```
+
+<!---END-->
+<inline-frame src="./resources/dropNASelector.html" width="100%" height="500px"></inline-frame>
+
+<!---FUN dropNASelectorSeveralCols-->
+
+```kotlin
 df.dropNA { age and weight } // remove rows where either 'age' or 'weight' is null or NaN
+```
+
+<!---END-->
+<inline-frame src="./resources/dropNASelectorSeveralCols.html" width="100%" height="500px"></inline-frame>
+
+<!---FUN dropNAWhereAllNASelector-->
+
+```kotlin
 df.dropNA(whereAllNA = true) { age and weight } // remove rows where both 'age' and 'weight' are null or NaN
 ```
 
-<inline-frame src="resources/org.jetbrains.kotlinx.dataframe.samples.api.Access.dropNA.html" width="100%"/>
 <!---END-->
+<inline-frame src="./resources/dropNAWhereAllNASelector.html" width="100%" height="500px"></inline-frame>
 
 If called on a [`DataColumn`](DataColumn.md), removes [`NA` values](nanAndNa.md#na) from this [`DataColumn`](DataColumn.md),
 adjusting the type accordingly.
