+++ samples/build.gradle.kts
@@ -117,6 +117,7 @@ korro {
                 include("groupBy.md")
                 include("pivot.md")
                 include("countDistinct.md")
+                include("drop.md")
                 include("filter.md")
                 include("count.md")
                 include("valueCounts.md")
